Fix wrong window function results for float `PARTITION BY` keys with several threads
