+++ tests/queries/0_stateless/05043_window_partition_by_float_scatter.reference
@@ -1,0 +1,4 @@
+100000
+100000
+0
+1
